refactor: IssueDetailContext 관심사 분리
