+++ .env
@@ -1,2 +1,2 @@
-REACT_APP_GITHUB_ACCESS_TOKEN = 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
+REACT_APP_GITHUB_ACCESS_TOKEN = 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
 REACT_APP_GITHUB_API_URL = https://api.github.com
