+++ README.md
@@ -429,9 +429,9 @@
 
   - Case 03.
 
-    - Context API
+    - Context API (IssueContext)
 
-      // 횡단 관심사에 대한 분리가 되어있지 않아서 Custom Hook과 각종 함수를 통해 관심사 분리를 하였습니다. 
+      횡단 관심사에 대한 분리가 되어있지 않아서 Custom Hook과 각종 함수를 통해 관심사 분리를 하였습니다. 
 
       ```ts
       // 깃허브에서 이슈 HTTP GET 요청
@@ -573,6 +573,10 @@
 
       Intersection Observer API 기반으로 무한스크롤을 구현하는 프로바이더 코드를 분리하여 관심사 분리를 하고, 횡단 관심사를 유지 할 수 있게 합니다. 참고로 하나의 파일 안에 모두 작성되어있습니다. 재사용에 대한 필요성을 느끼지 않아서 파일을 따로 만들지않고, 오히려 같이 나열하여 코드의 흐름을 한눈에 파악하여 유지보수를 하기 쉽게하려고 해당 코드를 작성하였습니다.
 
-      
+    - Context API (IssueDetailContext)
 
+      횡단 관심사에 대한 분리가 되어있지 않아서 Custom Hook을 통해 관심사 분리를 하였습니다. 
       
+      ```ts
+      
+      ```
